use inheritance for the display method
