+++ bin/main.rb
@@ -1,5 +1,7 @@
 require_relative '../lib/draw'
 
+require_relative '../lib/logic'
+
 puts 'Starting Tic Tac game'
 
 puts 'Enter player 1 name'
@@ -18,29 +20,23 @@ player2 = player2.capitalize
 
 puts ''
 
-
-
 sleep 1
 
 system 'cls'
 
 system 'clear'
 
-
-
 players = %w[O X]
 
 current_player = players[rand(2)]
 
-b = Board.new(current_player)
+b = Draw.new(current_player)
 
 puts
 
 puts "Welcome to the board \n #{player1} will play as #{players[0]} and #{player2} will play as #{players[1]}"
 
 puts "Let's start"
-
-
 
 sleep 2
 
@@ -50,7 +46,7 @@ system 'clear'
 
 b.display
 
-while not b.board_full and not b.winner
+while !b.board_full and !b.winner
 
   b.ask_player_for_move(current_player)
 
@@ -62,106 +58,14 @@ while not b.board_full and not b.winner
 
 end
 
-
-
-if b.winner()
+if b.winner
 
   puts "Player #{b.get_next_turn} wins"
 
 else
 
-  puts "The game is Tie"
+  puts 'The game is Tie'
 
-puts "Game over"
-
-
-
-
-
-# the_loop = 0..4
-
-# the_loop.each do
-
-#   b.display
-
-#   puts "it's #{player1}'s turn"
-
-#   puts 'Please select an available cell from the board'
-
-#   loop do
-
-#     input = gets.chomp
-
-#     case input.to_i
-
-#     when input.nil? || !(input.to_i >= 1 && input.to_i < 10)
-
-#       next
-
-#     when 1..9
-
-#       input.to_i
-
-#       puts " #{player1} value valid"
-
-#       break
-
-#     else puts 'invalid value, please select from the available cell only'
-
-#     end
-
-#   end
-
-
-
-#   system 'cls'
-
-#   system 'clear'
-
-
-
-#   b.display
-
-#   puts "it's #{player2}'s turn"
-
-#   puts 'Please select an available cell from the board'
-
-#   loop do
-
-#     input = gets.chomp
-
-#     case input.to_i
-
-#     when input.nil? || !(input.to_i >= 1 && input.to_i < 10)
-
-#       next
-
-#     when 1..9
-
-#       input.to_i
-
-#       puts "#{player2} value valid"
-
-#       break
-
-#     else puts 'invalid value, please select from the available cell only'
-
-#     end
-
-#   end
-
-#   system 'cls'
-
-#   system 'clear'
-
-# end
-
-
-
-# puts "It's a tie for now"
-
-# puts 'till we work on logic'
-
-# puts 'thank you'
+  puts 'Game over'
 
 end
